Refactoring (retiling is temporarily disabled)
The code is reworked so that the input image is read as a simple 2d
array without recognizing tiles. For "raw" images that are stored a
tile at a time, this means that the input should be described as having
the width of a single tile.

From here on, retiling will be treated as a separate step.