Config format changes
The TOML data is completely rearranged, and code adjusted to match.
The code no longer demands a `.bin` extension for the input (in
preparation for uses where the input isn't raw, or the output is), and
the config specifies an extension for the output (so imageio can
choose a corresponding output format).