Re-implement tiling
Added `get_tiles` and `arrange_tiles` functions which extract a grid of
tiles from the input (trimming any specified padding), re-work them
into a new grid 16 columns wide (for now), and make a new image from
that tile grid (re-padding between and around tiles as specified).

The `tiles.result.bg` config value specifies a constant bitmap value to
use for padding the tiles in the output (and adding a border). Bitmap
data from the input outside of the tiles is ignored - in particular, it
is NOT checked against the bg value (by design).