Work with colour channels
It's now generally assumed that the Numpy arrays representing images
will have a final dimension for the colour planes of the image data.

When a palettized image is read, the palette index values are looked up
in a 2d array (index x planes); thus, the colour-plane dimension is
added to the data.